+++ back-end/api/routes.js
@@ -1,7 +1,9 @@
 const user = require('../controllers/user');
 const products = require('../controllers/products');
+const login = require('../controllers/login');  
 
 module.exports = {
   user,
   products,
+  login,
 };
