+++ back-end/api/server.js
@@ -2,7 +2,8 @@ const bodyParser = require('body-parser');
 const express = require('express');
 const path = require('path');
 const cors = require('cors');
-const { user, products } = require('./routes');
+const verifyJWT = require('../middlewares/verifyJWT')
+const { user, products, login } = require('./routes');
 
 const app = express();
 
@@ -11,7 +12,8 @@ app.use(express.static(path.resolve(__dirname, '..', 'public', 'images')));
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 
-app.use(user);
-app.use(products);
+app.use('/user', user);
+app.use('/login', login)
+app.use('/products', verifyJWT, products);
 
 module.exports = app;
