+++ back-end/controllers/login.js
@@ -1,0 +1,19 @@
+const express = require('express');
+const rescue = require('../rescue');
+const login = require('../models/login');
+
+const router = express.Router();
+
+const login = async (req, res) => {
+  const { password, email } = req.body;
+  if (!email || !password)
+    return res.status(422).json({ message: 'Campos vazios!' });
+  const user = await login(email, password);
+  if (!user) return res.status(401).json({ message: 'Usuário não encontrado' });
+  const token = generateJWT(email);
+  res.status(200).json({ name: user.name, token, email, role: user.admin });
+};
+
+router.post('/', rescue(login));
+
+module.exports = router;
