+++ back-end/controllers/products.js
@@ -1,16 +1,12 @@
 const express = require('express');
-
-const router = express.Router();
-
 const rescue = require('../rescue');
-
 const Products = require('../models/products');
 
-// const verifyJWT = require('../middlewares/verifyJWT')
+const router = express.Router();
 
 const callBackAllProducts = (_req, res) =>
   Products.allProducts().then(body => res.status(201).json(body));
 
-router.get('/products', rescue(callBackAllProducts));
+router.get('/', rescue(callBackAllProducts));
 
 module.exports = router;
