+++ back-end/models/login.js
@@ -1,0 +1,11 @@
+const conn = require('../connection');
+
+module.exports = async function (email, password) {
+  const query = `SELECT * FROM user WHERE email = '${email}' AND password = '${password}'`;
+  return new Promise((resolve, reject) => {
+    conn.query(query, (err, results) => {
+      if (err) return reject(err);
+      return resolve(results[0]);
+    });
+  });
+};
